Force tcpdump to no longer buffer packets and flush to disk
To ensure packet data is available in the output file when
the process is killed, we should flush to disk far more often
than the default.  This patch adds the -U flag to tcpdump, which
will force a flush to disk on each packet rather than waiting for
the 4k buffer to fill.